Added MatOfDMatch

+++ src/OpenCvSharp.CPlusPlus/modules/core/Mat/MatOfDouble.cs
@@ -446,7 +446,7 @@ namespace OpenCvSharp.CPlusPlus
         /// <param name="s">An optional value to initialize each matrix element with. 
         /// To set all the matrix elements to the particular value after the construction, use SetTo(Scalar s) method .</param>
 #endif
-        public MatOfDouble(IEnumerable<int> sizes, int s)
+        public MatOfDouble(IEnumerable<int> sizes, double s)
             : base(sizes, ThisType, s)
         {
         }
